+++ tests/pyincore/models/test_networkdataset.py
@@ -1,6 +1,5 @@
 import json
 import os
-from pathlib import Path
 
 import pytest
 from networkx import Graph