+++ tests/test_format.py
@@ -62,6 +62,8 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/housingrecovery/housingrecovery.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/housingrecovery/housingrecoveryutil.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/housingrecovery/test_housingrecovery.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/pipelinefunctionality/pipelinefunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/pipelinefunctionality/test_pipelinefunctionality.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/utils/test_networkutil.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/models/test_networkdataset.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_analysis.py'),