+++ tests/pyincore/analyses/wfnfunctionality/test_wfnfunctionality.py
@@ -6,9 +6,9 @@ from pyincore import IncoreClient, FragilityService, MappingSet, NetworkDataset,
 from pyincore.analyses.montecarlofailureprobability import MonteCarloFailureProbability
 from pyincore.analyses.wfnfunctionality import WfnFunctionality
 from pyincore.analyses.pipelinedamagerepairrate import PipelineDamageRepairRate
+from pyincore.analyses.pipelinefunctionality import PipelineFunctionality
 from pyincore.analyses.waterfacilitydamage import WaterFacilityDamage
 
-import geopandas as gpd
 
 def run_with_base_class():
     client = IncoreClient()
@@ -79,16 +79,16 @@ def run_with_base_class():
 
     pipeline_dmg_w_rr.run_analysis()
 
-    pipeline_dmg_fs = pipeline_dmg_w_rr.get_output_dataset("result")
+    pipeline_dmg_w_rr_ds = pipeline_dmg_w_rr.get_output_dataset("result")
 
     # Using water facility damage, run a Monte Carlo analysis
-    print("Run Monte Carlo failure probability")
-    mc = MonteCarloFailureProbability(client)
+    nsamp = 20 #20000
 
+    print("Run Monte Carlo failure")
+    mc = MonteCarloFailureProbability(client)
     mc.set_input_dataset("damage", wf_dmg_result)
     mc.set_parameter("result_name", "wf_dmg_mc")
     mc.set_parameter("num_cpu", 8)
-    nsamp = 20000
     mc.set_parameter("num_samples", nsamp)
     mc.set_parameter("damage_interval_keys", ["DS_0", "DS_1", "DS_2", "DS_3", "DS_4"])
     mc.set_parameter("failure_state_keys", ["DS_3", "DS_4"])
@@ -97,13 +97,24 @@ def run_with_base_class():
 
     wf_dmg_fs = mc.get_output_dataset("sample_failure_state")
 
+    # Run pipeline functionality analysis
+    print("Run pipeline functionality")
+    pp_func = PipelineFunctionality(client)
+    pp_func.set_parameter("result_name", "mmsa_pipeline_functionality")
+    pp_func.set_parameter("num_samples", nsamp)
+    pp_func.set_input_dataset("pipeline_repair_rate_damage", pipeline_dmg_w_rr_ds)
+
+    pp_func.run_analysis()
+
+    pp_dmg_fs = pp_func.get_output_dataset("sample_failure_state")
+
     # Run wfn functionality
     print("Run water facility network functionality analysis")
     wfn_func = WfnFunctionality(client)
 
     wfn_func.load_remote_input_dataset("wfn_network", wfn_dataset_id)
     wfn_func.set_input_dataset("wf_sample_failure_state", wf_dmg_fs)
-    wfn_func.set_input_dataset("pp_sample_failure_state", pipeline_dmg_fs)
+    wfn_func.set_input_dataset("pp_sample_failure_state", pp_dmg_fs)
     wfn_func.set_parameter("result_name", "mmsa_wfn_functionality")
     wfn_func.set_parameter("tank_node_list", [1, 7, 10, 13, 14, 15])
     wfn_func.set_parameter("pumpstation_node_list", [2, 3, 4, 5, 6, 8, 9, 11, 12])