+++ tests/test_format.py
@@ -71,7 +71,9 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/waterfacilityrestoration/waterfacilityrestorationutil.py'),
     os.path.join(PYINCORE_ROOT_FOLDER,
                  'tests/pyincore/analyses/waterfacilityrestoration/test_waterfacilityrestoration.py'),
-    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/socialvulnerability/socialvulnerability.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/wfnfunctionality/wfnfunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/wfnfunctionality/test_wfnfunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/socialvulnerability/socialvulnverability.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/socialvulnerability/test_socialvulnerability.py')
 ]