+++ tests/test_format.py
@@ -61,7 +61,24 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/utils/test_popdisloutputprocess.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/housingrecovery/housingrecovery.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/housingrecovery/housingrecoveryutil.py'),
-    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/housingrecovery/test_housingrecovery.py')
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/housingrecovery/test_housingrecovery.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, "pyincore/analyses/epnfunctionality/epnfunctionality.py"),
+    os.path.join(PYINCORE_ROOT_FOLDER, "pyincore/analyses/epnfunctionality/epnfunctionalityutil.py"),
+    os.path.join(PYINCORE_ROOT_FOLDER, "tests/pyincore/analyses/epnfunctionality/test_epnfunctionality.py"),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/pipelinefunctionality/pipelinefunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/pipelinefunctionality/test_pipelinefunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/utils/test_networkutil.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/models/test_networkdataset.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_analysis.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/wfnfunctionality/wfnfunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/wfnfunctionality/wfnfunctionalityutil.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER,  'tests/pyincore/analyses/wfnfunctionality/test_wfnfunctionality.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/waterfacilityrestoration/waterfacilityrestoration.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/waterfacilityrestoration/waterfacilityrestorationutil.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER,
+                 'tests/pyincore/analyses/waterfacilityrestoration/test_waterfacilityrestoration.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/socialvulnerability/socialvulnerability.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/socialvulnerability/test_socialvulnerability.py')
 ]