+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [1.6.0] - Unreleased
+## [Unreleased]
 
 ### Added
 - Pipeline functionality analysis working with MMSA Shelby buried pipelines [#176](https://github.com/IN-CORE/pyincore/pull/176)
+
+### Changed
+- Network utils refactored to use network dataset [#149](https://github.com/IN-CORE/pyincore/issues/149)
 
 
 ## [1.5.0] - 2022-06-29